move all lights to their own folder. Add icons for all light types
