+++ source/client/components/lights/CVDirectionalLight.ts
@@ -15,26 +15,33 @@
  * limitations under the License.
  */
 
-import CAmbientLight from "@ff/scene/components/CAmbientLight";
+import CDirectionalLight from "@ff/scene/components/CDirectionalLight";
+import { Node } from "@ff/graph/Component";
 
 import { IDocument, INode, ILight, ColorRGB, TLightType } from "client/schema/document";
 
 import { ICVLight } from "./CVLight";
+import { EShadowMapResolution } from "@ff/scene/components/CLight";
+import CVNode from "../CVNode";
 
 ////////////////////////////////////////////////////////////////////////////////
 
-export default class CVAmbienLight extends CAmbientLight implements ICVLight
+export default class CVDirectionalLight extends CDirectionalLight implements ICVLight
 {
-    static readonly typeName: string = "CVAmbientLight";
-    static readonly type: TLightType = "ambient";
+    static readonly typeName: string = "CVDirectionalLight";
+    static readonly type: TLightType = "directional";
 
-    static readonly text: string = "Ambient Light";
-    static readonly icon: string = "bulb";
+    static readonly text: string = "Directional Light";
+    static readonly icon: string = "sun";
 
     get settingProperties() {
         return [
             this.ins.color,
             this.ins.intensity,
+            this.ins.shadowEnabled,
+            this.ins.shadowSize,
+            this.ins.shadowResolution,
+            this.ins.shadowBlur,
         ];
     }
 
@@ -46,6 +53,10 @@ export default class CVAmbienLight extends CAmbientLight implements ICVLight
     }
 
     dispose(): void {
+        if(this.ins.shadowEnabled.value && this.light.shadow.map) {
+            this.light.shadow.map.dispose();
+        }
+
         super.dispose()
     }
 
@@ -58,15 +69,21 @@ export default class CVAmbienLight extends CAmbientLight implements ICVLight
         const data = document.lights[node.light];
         const ins = this.ins;
 
-        if (data.type !== CVAmbienLight.type) {
-            throw new Error("light type mismatch: not an ambient light");
+        if (data.type !== "directional") {
+            throw new Error("light type mismatch: not a directional light");
         }
-
-        data.point = data.point || {} as any;
 
         ins.copyValues({
             color: data.color !== undefined ? data.color : ins.color.schema.preset,
             intensity: data.intensity !== undefined ? data.intensity : ins.intensity.schema.preset,
+
+            position: ins.position.schema.preset,
+            target: ins.target.schema.preset,
+
+            shadowEnabled: data.shadowEnabled || false,
+            shadowSize: data.shadowSize !== undefined ? data.shadowSize : ins.shadowSize.schema.preset,
+            shadowResolution: data.shadowResolution !== undefined ? EShadowMapResolution[data.shadowResolution] || 0 : ins.shadowResolution.schema.preset,
+            shadowBlur: data.shadowBlur !== undefined ? data.shadowBlur : ins.shadowBlur.schema.preset,
         });
 
         return node.light;
@@ -78,10 +95,24 @@ export default class CVAmbienLight extends CAmbientLight implements ICVLight
 
         const data = {
             color: ins.color.cloneValue() as ColorRGB,
-            intensity: ins.intensity.value,
+            intensity: ins.intensity.value
         } as ILight;
 
-        data.type = CVAmbienLight.type;
+        data.type = CVDirectionalLight.type;
+
+        if (ins.shadowEnabled.value) {
+            data.shadowEnabled = true;
+
+            if (!ins.shadowSize.isDefault()) {
+                data.shadowSize = ins.shadowSize.value;
+            }
+            if (!ins.shadowBlur.isDefault()) {
+                data.shadowBlur = ins.shadowBlur.value;
+            }
+            if (!ins.shadowResolution.isDefault()) {
+                data.shadowResolution = EShadowMapResolution[ins.shadowResolution.value];
+            }
+        }
 
         document.lights = document.lights || [];
         const lightIndex = document.lights.length;
