+++ source/client/components/lights/CVLight.ts
