+++ source/client/components/lights/CVPointLight.ts
