+++ source/client/ui/explorer/MainView.ts
@@ -48,6 +48,10 @@ Icon.add("zoom", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 5
 Icon.add("tools", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M501.1 395.7L384 278.6c-23.1-23.1-57.6-27.6-85.4-13.9L192 158.1V96L64 0 0 64l96 128h62.1l106.6 106.6c-13.6 27.8-9.2 62.3 13.9 85.4l117.1 117.1c14.6 14.6 38.2 14.6 52.7 0l52.7-52.7c14.5-14.6 14.5-38.2 0-52.7zM331.7 225c28.3 0 54.9 11 74.9 31l19.4 19.4c15.8-6.9 30.8-16.5 43.8-29.5 37.1-37.1 49.7-89.3 37.9-136.7-2.2-9-13.5-12.1-20.1-5.5l-74.4 74.4-67.9-11.3L334 98.9l74.4-74.4c6.6-6.6 3.4-17.9-5.7-20.2-47.4-11.7-99.6.9-136.6 37.9-28.5 28.5-41.9 66.1-41.2 103.6l82.1 82.1c8.1-1.9 16.5-2.9 24.7-2.9zm-103.9 82l-56.7-56.7L18.7 402.8c-25 25-25 65.5 0 90.5s65.5 25 90.5 0l123.6-123.6c-7.6-19.9-9.9-41.6-5-62.7zM64 472c-13.2 0-24-10.8-24-24 0-13.3 10.7-24 24-24s24 10.7 24 24c0 13.2-10.7 24-24 24z"/></svg>`);
 Icon.add("environment", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 256 256"><path d="M219.81,14.75H36.19A21.19,21.19,0,0,0,15,35.94V220.06a21.19,21.19,0,0,0,21.19,21.19H219.81A21.19,21.19,0,0,0,241,220.06V35.94A21.19,21.19,0,0,0,219.81,14.75Zm-2.65,205.31H38.84a2.64,2.64,0,0,1-2.65-2.65V38.59a2.64,2.64,0,0,1,2.65-2.65H217.16a2.64,2.64,0,0,1,2.65,2.65V217.41A2.64,2.64,0,0,1,217.16,220.06ZM72.16,71.59A18.32,18.32,0,1,0,90.47,89.91,18.32,18.32,0,0,0,72.16,71.59ZM57.51,198.15H204V141.53l-40.06-40.06a5.49,5.49,0,0,0-7.77,0l-54.72,54.71L83.36,138.09a5.47,5.47,0,0,0-7.76,0L57.51,156.18Z"/></svg>`);
 Icon.add("bulb", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 512"><path d="M44.73 323.21c-7.65 4.42-10.28 14.2-5.86 21.86l8 13.86c4.42 7.65 14.21 10.28 21.86 5.86l93.26-53.84a207.865 207.865 0 0 1-26.83-39.93l-90.43 52.19zM112.46 168H16c-8.84 0-16 7.16-16 16v16c0 8.84 7.16 16 16 16h100.21a210.423 210.423 0 0 1-3.75-48zm127.6 291.17c0 3.15.93 6.22 2.68 8.84l24.51 36.84c2.97 4.46 7.97 7.14 13.32 7.14h78.85c5.36 0 10.36-2.68 13.32-7.14l24.51-36.84c1.74-2.62 2.67-5.7 2.68-8.84l.05-43.18H240.02l.04 43.18zM44.73 60.78l78.98 45.6c5.37-15.29 12.97-29.48 21.64-42.93L68.73 19.21c-7.65-4.42-17.44-1.8-21.86 5.86l-8 13.86c-4.42 7.65-1.79 17.44 5.86 21.85zm550.54 0c7.65-4.42 10.28-14.2 5.86-21.86l-8-13.86c-4.42-7.65-14.21-10.28-21.86-5.86l-76.61 44.23c8.68 13.41 15.76 27.9 21.2 43.19l79.41-45.84zm0 262.43l-90.97-52.52c-7.33 14.23-15.8 27.88-26.36 40.21l93.33 53.88c7.65 4.42 17.44 1.8 21.86-5.86l8-13.86c4.42-7.64 1.79-17.43-5.86-21.85zM624 168h-96.41c.1 2.68.41 5.3.41 8 0 13.54-1.55 26.89-4.12 40H624c8.84 0 16-7.16 16-16v-16c0-8.84-7.16-16-16-16zM320 80c-52.94 0-96 43.06-96 96 0 8.84 7.16 16 16 16s16-7.16 16-16c0-35.3 28.72-64 64-64 8.84 0 16-7.16 16-16s-7.16-16-16-16zm0-80C217.72 0 144 82.97 144 176c0 44.37 16.45 84.85 43.56 115.78 16.64 18.99 42.74 58.8 52.42 92.16v.06h48v-.12c-.01-4.77-.72-9.51-2.15-14.07-5.59-17.81-22.82-64.77-62.17-109.67-20.53-23.43-31.52-53.14-31.61-84.14-.2-73.64 59.67-128 127.95-128 70.58 0 128 57.42 128 128 0 30.97-11.24 60.85-31.65 84.14-39.11 44.61-56.42 91.47-62.1 109.46a47.507 47.507 0 0 0-2.22 14.3v.1h48v-.05c9.68-33.37 35.78-73.18 52.42-92.16C479.55 260.85 496 220.37 496 176 496 78.8 417.2 0 320 0z"/></svg>`);
+Icon.add("spot", html`<svg viewBox="0 0 512 512" xmlns="http://www.w3.org/2000/svg"><path d="M301.87 108.727c0 52.94 43.06 96 96 96 8.84 0 16-7.16 16-16s-7.16-16-16-16c-35.3 0-64-28.72-64-64 0-8.84-7.16-16-16-16s-16 7.16-16 16zM478.636 144.265a15.898 15.898 0 0 0 4.356-8.146l8.718-43.38a16.008 16.008 0 0 0-4.37-14.468l-55.755-55.755a15.982 15.982 0 0 0-14.467-4.37l-43.381 8.718a16.084 16.084 0 0 0-8.146 4.356l-30.568 30.498L448.13 174.826zM30.3 216.774c-7.856 4.045-10.956 13.686-6.912 21.55l7.32 14.232c4.044 7.855 13.695 10.956 21.55 6.911l207.354-102.82c-5.12-15.272-4.739-35.452-3.706-49.874zM257.42 455.585c-3.858 7.948-.554 17.522 7.4 21.388l14.393 6.994c7.949 3.858 17.53.549 21.388-7.4l108.397-220.653c-14.872 2.094-34.317 1.858-49.625-2.162zM119.554 363.004c-6.162 6.331-6.037 16.458.298 22.63l11.464 11.165c6.331 6.162 16.468 6.034 22.63-.297L309.885 230.5c-11.419-7.558-22.03-17.993-30.532-31.249z"/></svg>`);
+Icon.add("area", html`<svg viewBox="0 0 640 512" xmlns="http://www.w3.org/2000/svg"><path style="stroke-linecap:round;-inkscape-stroke:none" d="M351.053 207.98a25 25 0 0 0-12.014 4.811l-7.65 5.729a25 25 0 0 0-5.028 34.996 25 25 0 0 0 34.996 5.027l7.65-5.729a25 25 0 0 0 5.028-34.996 25 25 0 0 0-22.982-9.838zm-58.348 43.956a25 25 0 0 0-11.848 5.203L45.99 445.346a25 25 0 0 0-3.877 35.14 25 25 0 0 0 35.143 3.877l234.867-188.207a25 25 0 0 0 3.875-35.142 25 25 0 0 0-23.293-9.078zM131.023 68.966c-7.535 0-12.006 1.996-17.07 6.53-5.064 4.535-11.212 17.475-6.45 28.72a25.003 25.003 0 0 0 .017.036l114 267.99a25.003 25.003 0 0 0 .048.115c.146.338.588 2.384 3.852 6.344 3.264 3.96 12.068 9.254 19.603 9.254h335.891c7.51 0 12.01-1.991 17.082-6.539 5.073-4.548 11.19-17.488 6.438-28.709a25.003 25.003 0 0 0-.014-.037l-114-267.99a25.003 25.003 0 0 0-.018-.04c-.155-.363-.602-2.438-3.883-6.42-3.28-3.98-12.096-9.253-19.605-9.253H131.025zM444.428 104.3l.677 1.59c-.513-.992-.653-1.531-.677-1.59zm-276.313 14.668h282.553l93.154 218.988H261.27zm98.705 232.037c.48.923.597 1.378.623 1.467z"/></svg>`);
+Icon.add("sun", html`<svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg"><path style="stroke:none" d="m10.824 3.745 1.711-1.705c.425-.416 1.82 1.005 1.408 1.42l-1.707 1.702zm2.18 3.247 2.416.004c.595.006.576 1.997-.008 2l-2.41-.004zm-.765 3.833 1.704 1.711c.416.425-1.005 1.82-1.42 1.408l-1.701-1.708zM8.997 13l-.004 2.416c-.006.594-1.997.576-2-.01l.004-2.41zm-3.825-.758-1.72 1.697c-.426.414-1.814-1.013-1.4-1.427l1.715-1.693zM2.992 8.99.578 8.986c-.594-.006-.576-1.997.009-2l2.41.004zm.76-3.818L2.045 3.46c-.416-.424 1.005-1.819 1.42-1.408L5.168 3.76zM11 8a3 3 0 1 1-6 0 3 3 0 0 1 6 0zM7 3V.584C7.005-.01 8.996.004 9 .59V3z"/></svg>`);
+Icon.add("half-sun", html`<svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg"><path style=";stroke-width:50;stroke-linecap:round;stroke-linejoin:round" d="M.588 9.16h14.84c.212 0 .384.172.384.385v1.124a.384.384 0 0 1-.385.385H.588a.384.384 0 0 1-.384-.385V9.545c0-.213.171-.385.384-.385zM8 .145c-.499-.003-.998.142-1 .439V3h2V.59C8.998.297 8.499.147 8 .145zm4.734 1.82a.274.274 0 0 0-.199.074l-1.71 1.707 1.411 1.416 1.707-1.701c.36-.364-.663-1.497-1.209-1.496zm-9.464.012C2.725 1.974 1.68 3.09 2.045 3.46l1.707 1.71L5.168 3.76 3.467 2.053a.272.272 0 0 0-.197-.076zM8 5a2.997 2.997 0 0 0-2.988 2.785h5.976A2.997 2.997 0 0 0 8 5zM.586 6.986c-.253.002-.4.375-.438.8h2.846l.002-.796zm12.418.006v.793h2.848c-.034-.419-.177-.786-.432-.789z"/></svg>`);
 Icon.add("tape", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 512"><path d="M624 432H362.3c52.1-41 85.7-104.5 85.7-176 0-123.7-100.3-224-224-224S0 132.3 0 256s100.3 224 224 224h400c8.8 0 16-7.2 16-16v-16c0-8.8-7.2-16-16-16zm-400 0c-97 0-176-79-176-176S127 80 224 80s176 79 176 176-79 176-176 176zm0-272c-53 0-96 43-96 96s43 96 96 96 96-43 96-96-43-96-96-96zm0 144c-26.5 0-48-21.5-48-48s21.5-48 48-48 48 21.5 48 48-21.5 48-48 48z"/></svg>`);
 Icon.add("knife", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512"><path d="M566.28 88.57c12.96-12.5 12.96-32.76 0-45.25L531.07 9.37c-12.96-12.5-33.98-12.5-46.94 0L319.99 160 4.76 464.14c-8.25 7.96-5.38 22.16 5.53 25.69C53.72 503.86 102.37 512 150.51 512c75.83 0 150.42-20.19 201.49-69.35l104.4-100.04c12.95-12.41 13.17-33.05.49-45.73L448 288v-80L566.28 88.57zM496 64c8.84 0 16 7.16 16 16s-7.16 16-16 16-16-7.16-16-16 7.16-16 16-16zM318.71 408.07C281.24 444.14 221.5 464 150.51 464c-23.16 0-46.79-2.1-70.07-6.17L319.4 227.28l91.99 91.99-92.68 88.8zM432 160c-8.84 0-16-7.16-16-16s7.16-16 16-16 16 7.16 16 16-7.17 16-16 16z"/></svg>`);
 Icon.add("bars", html`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M16 132h416c8.837 0 16-7.163 16-16V76c0-8.837-7.163-16-16-16H16C7.163 60 0 67.163 0 76v40c0 8.837 7.163 16 16 16zm0 160h416c8.837 0 16-7.163 16-16v-40c0-8.837-7.163-16-16-16H16c-8.837 0-16 7.163-16 16v40c0 8.837 7.163 16 16 16zm0 160h416c8.837 0 16-7.163 16-16v-40c0-8.837-7.163-16-16-16H16c-8.837 0-16 7.163-16 16v40c0 8.837 7.163 16 16 16z"/></svg>`);
